amos/image_labelling_service/requirements.txt
django
djangorestframework
django-swagger-render
google-cloud-vision
Image
mock
requests
psycopg2-binary
pytest
pytest-cov